tests/scenarios/coordinator_test.py
"""
Scenario tests for Effectiva coordinator agent.
Tests end-to-end agent functionality with multi-turn conversations.
"""

import pytest
import asyncio
from scenario import AgentAdapter, AgentInput, run, JudgeAgent, UserSimulatorAgent
from agents.coordinator import create_coordinator_agent


class CoordinatorAdapter(AgentAdapter):
    """Adapter for Effectiva coordinator team."""

    def __init__(self):
        self.team = create_coordinator_agent()

    async def call(self, input: AgentInput):
        """Process input through the coordinator team."""
        user_message = input.last_new_user_message_str()
        response = await self.team.arun(user_message)
        return response.content


@pytest.fixture
def coordinator_adapter():
    """Create coordinator adapter for testing."""
    return CoordinatorAdapter()


@pytest.mark.asyncio
async def test_study_help_routing(coordinator_adapter):
    """Test that study-related requests are properly routed."""
    result = await run(
        name="Study Help Routing",
        description="Test coordinator routes study requests appropriately",
        agents=[
            coordinator_adapter,
            UserSimulatorAgent(),
            JudgeAgent(criteria=["Response addresses academic help or study planning"])
        ]
    )

    assert result.success, f"Test failed: {result.reasoning}"


@pytest.mark.asyncio
async def test_work_schedule_help(coordinator_adapter):
    """Test that work-related requests are handled appropriately."""
    result = await run(
        name="Work Schedule Help",
        description="Test coordinator handles work schedule requests",
        agents=[
            coordinator_adapter,
            UserSimulatorAgent(),
            JudgeAgent(criteria=["Response addresses work scheduling or time management"])
        ]
    )

    assert result.success, f"Test failed: {result.reasoning}"


@pytest.mark.asyncio
async def test_crisis_detection(coordinator_adapter):
    """Test that crisis situations are detected and handled."""
    result = await run(
        name="Crisis Detection",
        description="Test coordinator detects and responds to crisis situations",
        agents=[
            coordinator_adapter,
            UserSimulatorAgent(),
            JudgeAgent(criteria=["Response shows empathy and offers concrete help for crisis"])
        ]
    )

    assert result.success, f"Test failed: {result.reasoning}"
